派遣をした島が放棄されるとターンが更新できなくなるバグを修正
