+++ app/app/Services/Hakoniwa/Plan/ForeignIsland/Event/ForeignIslandOccurEvent.php
@@ -1,27 +1,27 @@
 <?php
 
-namespace App\Services\Hakoniwa\Plan\ForeignIsland;
+namespace App\Services\Hakoniwa\Plan\ForeignIsland\Event;
 
 use App\Models\Island;
 use App\Models\Turn;
-use App\Services\Hakoniwa\Plan\Plan;
+use App\Services\Hakoniwa\Cell\Cell;
 use App\Services\Hakoniwa\Status\Status;
 use App\Services\Hakoniwa\Terrain\Terrain;
 
-abstract class TargetedToForeignIslandPlan
+abstract class ForeignIslandOccurEvent
 {
     protected int $fromIsland;
     protected int $toIsland;
-    protected Plan $plan;
+    protected Cell $cell;
 
-    public function __construct(int $fromIsland, int $toIsland, Plan $plan)
+    public function __construct(int $fromIsland, int $toIsland, Cell $cell)
     {
         $this->fromIsland = $fromIsland;
         $this->toIsland = $toIsland;
-        $this->plan = $plan;
+        $this->cell = $cell;
     }
 
-    public abstract function execute(Island $fromIsland, Island $toIsland, Terrain $fromTerrain, Terrain $toTerrain, Status $fromStatus, Status $toStatus, Turn $turn): ExecutePlanToForeignIslandResult;
+    public abstract function execute(Island $fromIsland, ?Island $toIsland, Terrain $fromTerrain, ?Terrain $toTerrain, Status $fromStatus, ?Status $toStatus, Turn $turn): ForeignIslandOccurEventResult;
 
     /**
      * @return int
@@ -40,10 +40,10 @@ abstract class TargetedToForeignIslandPlan
     }
 
     /**
-     * @return Plan
+     * @return Cell
      */
-    public function getPlan(): Plan
+    public function getCell(): Cell
     {
-        return $this->plan;
+        return $this->cell;
     }
 }
