+++ app/app/Services/Hakoniwa/Plan/ForeignIsland/Plan/FundsTransportToForeignIslandPlan.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\Hakoniwa\Plan\ForeignIsland;
+namespace App\Services\Hakoniwa\Plan\ForeignIsland\Plan;
 
 use App\Models\Island;
 use App\Models\Turn;
@@ -8,13 +8,13 @@ use App\Services\Hakoniwa\Cell\Cell;
 use App\Services\Hakoniwa\Cell\Sea;
 use App\Services\Hakoniwa\Cell\Shallow;
 use App\Services\Hakoniwa\Cell\Ship\TransportShip;
-use App\Services\Hakoniwa\Log\ResourcesTransportationLog;
+use App\Services\Hakoniwa\Log\FundsTransportationLog;
 use App\Services\Hakoniwa\Log\Logs;
-use App\Services\Hakoniwa\Plan\ResourcesTransportationPlan;
+use App\Services\Hakoniwa\Plan\FundsTransportationPlan;
 use App\Services\Hakoniwa\Status\Status;
 use App\Services\Hakoniwa\Terrain\Terrain;
 
-class ResourcesTransportToForeignIslandPlan extends TargetedToForeignIslandPlan
+class FundsTransportToForeignIslandPlan extends TargetedToForeignIslandPlan
 {
     public function execute(Island $fromIsland, Island $toIsland, Terrain $fromTerrain, Terrain $toTerrain, Status $fromStatus, Status $toStatus, Turn $turn): ExecutePlanToForeignIslandResult
     {
@@ -26,9 +26,9 @@ class ResourcesTransportToForeignIslandPlan extends TargetedToForeignIslandPlan
             return in_array($cell::TYPE, [Sea::TYPE, Shallow::TYPE]);
         });
 
-        $amount = $this->plan->getAmount() * ResourcesTransportationPlan::UNIT;
+        $amount = $this->plan->getAmount() * FundsTransportationPlan::UNIT;
 
-        $toStatus->setResources($toStatus->getResources() + $amount);
+        $toStatus->setFunds($toStatus->getFunds() + $amount);
 
         if ($seaCells->count() >= 1) {
             /** @var Cell $seaCell */
@@ -36,8 +36,8 @@ class ResourcesTransportToForeignIslandPlan extends TargetedToForeignIslandPlan
             $toTerrain->setCell($seaCell->getPoint(), new TransportShip(point: $seaCell->getPoint(), elevation: $seaCell->getElevation()));
         }
 
-        $fromLogs->add(new ResourcesTransportationLog($toIsland, $turn, $amount, true));
-        $toLogs->add(new ResourcesTransportationLog($fromIsland, $turn, $amount, false));
+        $fromLogs->add(new FundsTransportationLog($toIsland, $turn, $amount, true));
+        $toLogs->add(new FundsTransportationLog($fromIsland, $turn, $amount, false));
 
         return new ExecutePlanToForeignIslandResult($fromTerrain, $toTerrain, $fromStatus, $toStatus, $fromLogs, $toLogs);
     }
