+++ app/app/Services/Hakoniwa/Plan/ForeignIsland/Plan/TargetedToForeignIslandPlan.php
@@ -1,28 +1,27 @@
 <?php
 
-namespace App\Services\Hakoniwa\Plan\ForeignIsland;
+namespace App\Services\Hakoniwa\Plan\ForeignIsland\Plan;
 
 use App\Models\Island;
 use App\Models\Turn;
-use App\Services\Hakoniwa\Cell\Cell;
 use App\Services\Hakoniwa\Plan\Plan;
 use App\Services\Hakoniwa\Status\Status;
 use App\Services\Hakoniwa\Terrain\Terrain;
 
-abstract class ForeignIslandOccurEvent
+abstract class TargetedToForeignIslandPlan
 {
     protected int $fromIsland;
     protected int $toIsland;
-    protected Cell $cell;
+    protected Plan $plan;
 
-    public function __construct(int $fromIsland, int $toIsland, Cell $cell)
+    public function __construct(int $fromIsland, int $toIsland, Plan $plan)
     {
         $this->fromIsland = $fromIsland;
         $this->toIsland = $toIsland;
-        $this->cell = $cell;
+        $this->plan = $plan;
     }
 
-    public abstract function execute(Island $fromIsland, ?Island $toIsland, Terrain $fromTerrain, ?Terrain $toTerrain, Status $fromStatus, ?Status $toStatus, Turn $turn): ExecutePlanToForeignIslandResult;
+    public abstract function execute(Island $fromIsland, Island $toIsland, Terrain $fromTerrain, Terrain $toTerrain, Status $fromStatus, Status $toStatus, Turn $turn): ExecutePlanToForeignIslandResult;
 
     /**
      * @return int
@@ -41,10 +40,10 @@ abstract class ForeignIslandOccurEvent
     }
 
     /**
-     * @return Cell
+     * @return Plan
      */
-    public function getCell(): Cell
+    public function getPlan(): Plan
     {
-        return $this->cell;
+        return $this->plan;
     }
 }
